requirements.txt
amqp==5.0.9
asgiref==3.5.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.1.7
django-filter==21.1
djangorestframework==3.13.1
greenlet==1.1.2
importlib-metadata==4.10.1
kombu==5.2.3
Markdown==3.3.6
mysqlclient==2.1.0
prompt-toolkit==3.0.24
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.31
sqlparse==0.4.2
typing_extensions==4.0.1
uWSGI==2.0.18
vine==5.0.0
wcwidth==0.2.5
zipp==3.7.0
